return errors from SendMessage instead of silently dropping them
